GG-36898 moved opencsv dependency to ignite-indexing module
